1.use loop in bvh building 2.use reflectance to determine photon behavior in pm/ppm 3.fix ray tracing in ppm

+++ src/integrator/ppm.h
@@ -23,8 +23,8 @@ struct Hitpoint {
     Spectrum energyWeight;
     int photonCount[3];
     
-    float x; // screen x
-    float y; // screen y
+    int x; // screen x
+    int y; // screen y
 
     Hitpoint() {
     }
